requirements.txt
yutube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib